Move recommendations card component to own directory
